[FE][SYNC-240] feat: new home!
