+++ src/views/Home.vue
@@ -1,367 +1,188 @@
 <template>
-  <b-container fluid>
-    <category-bar />
-    <div class="d-flex justify-content-center">
-      <HomeBanner
-        :title="focus.title"
-        :tags="focus.tags.slice(0, 4)"
-        :last-updated-at="focus.lastUpdatedAt"
-        :blocks="focus.blocks"
-        :news-id="focus._id"
-      />
-    </div>
-    <div v-for="(section, sectionIndex) in allArticles" :key="sectionIndex">
-      <b-row class="py-5 heading-bar">
-        <b-col class="d-flex justify-content-center align-items-center">
-          <SyncIcon :icon="section.iconPath" class="mr-3" />
-          <h1 class="home-heading">{{ section.title }}</h1>
-        </b-col>
-      </b-row>
-      <div class="d-flex justify-content-center">
-        <b-row
-          cols="1"
-          :cols-md="Math.min(2, section.content.length)"
-          :cols-lg="Math.min(3, section.content.length)"
-          style="max-width: 1024px"
-        >
+  <div>
+    <div v-if="doneInit" class="main-container">
+      <div class="container_block">
+        <div class="section_title">
+          焦點內容
+        </div>
+        <HomeBanner
+          :title="articleBanner.title"
+          :tags="articleBanner.tags.slice(0, 4)"
+          :last-updated-at="articleBanner.lastUpdatedAt"
+          :blocks="articleBanner.blocks"
+          :news-id="articleBanner._id"
+        />
+        <div class="gallery">
           <ArticleCard
-            v-for="(news, newsIndex) in section.content"
-            :key="newsIndex"
-            :category="news.category"
-            :title="news.title"
-            :views-count="news.viewsCount"
-            :tags="news.tags.slice(0, 2)"
-            :last-updated-at="news.lastUpdatedAt"
-            :edited-count="news.editedCount"
-            :blocks="news.blocks"
-            :article-id="news._id"
+            v-for="(article, index) in articlesLatest"
+            :key="index"
+            :category="article.category"
+            :title="article.title"
+            :views-count="article.viewsCount"
+            :tags="article.tags.slice(0, 2)"
+            :blocks="article.blocks"
+            :article-id="article._id"
+            :lg="lg"
           />
-        </b-row>
+        </div>
       </div>
-      <b-row>
-        <b-col class="d-flex justify-content-center mt-4 mb-5" no-gutters>
-          <b-button squared variant="outline-blue">閱讀更多</b-button>
-        </b-col>
-      </b-row>
+
+      <div class="container_block">
+        <div class="section_title">等待編輯</div>
+        <div
+          class="gallery"
+          :class="{ 'gallery__wider': lg }"
+        >
+          <AwaitEditCard
+            v-for="(article, index) in articlesLatest"
+            :key="index"
+            :category="article.category"
+            :title="article.title"
+            :views-count="article.viewsCount"
+            :tags="article.tags.slice(0, 2)"
+            :last-updated-at="article.lastUpdatedAt"
+            :blocks="article.blocks"
+            :article-id="article._id"
+            :lg="lg"
+          />
+        </div>
+      </div>
+
+      <div class="container_block">
+        <div class="section_title">熱門內容</div>
+        <div class="gallery">
+          <ArticleCard
+            v-for="(article, index) in articlesHot"
+            :key="index"
+            :category="article.category"
+            :title="article.title"
+            :views-count="article.viewsCount"
+            :tags="article.tags.slice(0, 2)"
+            :blocks="article.blocks"
+            :article-id="article._id"
+            :lg="lg"
+          />
+        </div>
+      </div>
     </div>
     <RecommendHashtag />
-
-  </b-container>
+  </div>
 </template>
 
 <script>
-import ArticleCard from '@/components/ArticleCard.vue'
-import CategoryBar from '@/components/CategoryBar.vue'
-import RecommendHashtag from '@/components/RecommendHashtag.vue'
 import articleAPI from '@/api/article'
-import { mapGetters } from 'vuex'
+import ArticleCard from '@/components/ArticleCard.vue'
+import AwaitEditCard from '@/components/AwaitEditCard.vue'
 import HomeBanner from '@/components/HomeBanner.vue'
+import RecommendHashtag from '@/components/RecommendHashtag.vue'
+
+function makeArticle(article = {}) {
+  return {
+    _id: article._id ?? '',
+    title: article.title ?? '',
+    tags: article.tags ?? [],
+    lastUpdatedAt: article.lastUpdatedAt ?? '',
+    blocks: article.blocks ?? [],
+    editedCount: article.editedCount ?? 0,
+    category: article.category ?? ''
+  }
+}
 
 export default {
-  name: 'Home',
+  name: 'NewHome',
   components: {
     ArticleCard,
-    HeadlineCard,
-    CategoryBar,
-    RecommendHashtag,
-    HomeBanner
+    AwaitEditCard,
+    HomeBanner,
+    RecommendHashtag
   },
   data() {
     return {
-      categoryList: ['政經', '國際', '社會', '科技', '環境', '生活', '運動'],
-      allArticles: [],
-      latestList: [],
-      exploreList: [],
-      hotList: [],
-      headline: [],
-      iconPaths: ['latest', 'hot', 'explore'],
-      pickedHeadline: 0,
-      headlineTimer: null,
-      HEADLINEINTERVAL: 5000,
-      focus: {}
+      articleBanner: {},
+      articlesLatest: [],
+      articlesHot: [],
+      windowWidth: window.innerWidth,
+      doneInit: false
     }
   },
   computed: {
-    ...mapGetters([
-      'isLogin'
-    ])
-  },
-  watch: {
-    '$route.query.category'() {
-      if (this.$route.query.category) {
-        this.getCategoryArticles(this.$route.query.category)
-      } else this.initializeHomepage()
+    lg() {
+      return this.windowWidth >= 1280
     }
   },
   created() {
-    if (this.$route.query.category) {
-      this.getCategoryArticles(this.$route.query.category)
-    } else this.initializeHomepage()
+    this.getArticle()
   },
-  beforeDestroy() {
-    clearInterval(this.headlineTimer)
+  mounted() {
+    this.$nextTick(() => {
+      window.addEventListener('resize', () => {
+        this.windowWidth = window.innerWidth
+        console.log('size changed')
+      })
+    })
   },
   methods: {
-    async initializeHomepage() {
-      await articleAPI
-        .get()
-        .then((response) => {
-          const { data } = response
-          if (data.code === 200) {
-            const articles = data.data[0].sort(
-              (a, b) => new Date(b.lastUpdatedAt) - new Date(a.lastUpdatedAt)
-            )
-            const latestArticles = {
-              content: []
-            }
-            articles.forEach((article) => {
-              const {
-                category,
-                _id,
-                title,
-                viewsCount,
-                tags,
-                lastUpdatedAt,
-                editedCount,
-                blocks
-              } = article
-              latestArticles.content.push({
-                _id,
-                category,
-                title,
-                viewsCount,
-                tags,
-                lastUpdatedAt,
-                editedCount,
-                blocks
-              })
-            })
-            const articlesHot = data.data[0].sort(
-              (a, b) => b.viewsCount - a.viewsCount
-            )
-            this.latestList = latestArticles.content.slice(0, 6)
-            this.hotList = articlesHot.slice(0, 6)
-            this.exploreList = this.getRandomArticles(
-              latestArticles.content,
-              Math.min(6, latestArticles.content.length)
-            )
-            this.headline = this.exploreList.slice(0, 5)
-            this.focus = this.exploreList[0]
-            this.allArticles = [
-              {
-                title: '最新同步',
-                content: this.latestList,
-                iconPath: this.iconPaths[0]
-              },
-              {
-                title: '熱門同步',
-                content: this.hotList,
-                iconPath: this.iconPaths[1]
-              },
-              {
-                title: '探索其他',
-                content: this.exploreList,
-                iconPath: this.iconPaths[2]
-              }
-            ]
-          }
-        })
-        .catch((err) => console.error(err))
-      this.resetTimer()
-    },
-    getRandomArticles(arr, n) {
-      var result = new Array(n)
-      var len = arr.length
-      var taken = new Array(len)
-      if (n > len) {
-        throw new RangeError(
-          'getRandomArticles: more elements taken than available'
-        )
+    async getArticle() {
+      try {
+        const { data } = await articleAPI.get()
+        if (data.code !== 200) return
+
+        const articleSortedLatest = data.data[0].sort((a, b) => new Date(b.lastUpdatedAt) - new Date(a.lastUpdatedAt))
+        this.articlesLatest = articleSortedLatest.slice(0, 6).map((article) => makeArticle(article))
+
+        const articleSortedHot = data.data[1].sort((a, b) => b.viewsCount - a.viewsCount)
+        this.articlesHot = articleSortedHot.slice(1, 7).map((article) => makeArticle(article))
+
+        this.articleBanner = makeArticle(articleSortedHot[0])
+      } catch (e) {
+        console.error(e)
+      } finally {
+        this.doneInit = true
       }
-      while (n--) {
-        var x = Math.floor(Math.random() * len)
-        result[n] = arr[x in taken ? taken[x] : x]
-        taken[x] = --len in taken ? taken[len] : len
-      }
-      return result
-    },
-    async getCategoryArticles(param) {
-      if (param) {
-        try {
-          const categoricNews = []
-          const { data } = await articleAPI.search(undefined, undefined, param)
-          const type = data.type
-          if (type === 'success') {
-            const articles = data.data.sort(
-              (a, b) => new Date(b.lastUpdatedAt) - new Date(a.lastUpdatedAt)
-            )
-            articles.forEach((article) => {
-              const {
-                category,
-                _id,
-                title,
-                viewsCount,
-                tags,
-                lastUpdatedAt,
-                editedCount,
-                blocks
-              } = article
-              categoricNews.push({
-                category,
-                _id,
-                title,
-                viewsCount,
-                tags,
-                lastUpdatedAt,
-                editedCount,
-                blocks
-              })
-            })
-            this.allArticles = [
-              {
-                title: param + '同步',
-                content: categoricNews,
-                iconPath: this.iconPaths[0]
-              }
-            ]
-            this.headline = this.getRandomArticles(
-              categoricNews,
-              Math.min(5, categoricNews.length)
-            )
-            this.resetTimer()
-          } else {
-            throw new Error(data.message)
-          }
-        } catch (error) {
-          console.error(error.message)
-          this.$bvModal.msgBoxOk(error.message)
-          this.$router.push({ path: '/' })
-        }
-      }
-    },
-    resetTimer() {
-      if (this.headlineTimer !== null) clearInterval(this.headlineTimer)
-      this.headlineTimer = setInterval(() => {
-        this.pickedHeadline += 1
-        this.pickedHeadline %= Math.min(this.headline.length, 5)
-      }, this.HEADLINEINTERVAL)
-      this.pickedHeadline = 0
     }
   }
 }
 </script>
 
-<style lang="scss" scoped>
+<style lang="scss">
 @import '@/assets/scss/news.scss';
-// .category-navbar{
-//   border-bottom: 1px solid $gray-light;
-// }
 
-// .category-item{
-//   font-size: 20px;
-//   line-height: 30px;
-//   letter-spacing: 8px;
-//   color: #232323;
-//   &:hover{
-//     color: $blue !important;
-//   }
-// }
+.main-container {
+  display: flex;
+  flex-direction: column;
+  max-width: 1240px;
+  margin: 4rem auto;
+  padding: 0 1.25rem;
+  box-sizing: content-box;
+  gap: 4rem;
 
-.home-heading {
-  font-style: normal;
-  font-weight: bold;
-  font-size: 32px;
-  line-height: 47px;
-  letter-spacing: 0.5em;
-  color: #000000;
-  margin: 0;
-}
-
-.btn-outline-blue {
-  color: $blue;
-  border: 2px solid $blue;
-  font-size: 20px;
-  width: 130px;
-  letter-spacing: 0.2em;
-  text-align: center;
-  text-indent: 0.2em;
-}
-
-.btn-outline-blue:active,
-.btn-outline-blue:hover {
-  color: $white;
-  background: $blue;
-  border: 2px solid $blue;
-  font-size: 20px;
-}
-
-.heading-bar {
-  border-top: 1px solid $gray-light;
-}
-
-.headline-pages {
-  text-align: center;
-  margin-top: 38px;
-  margin-bottom: 64px;
-}
-
-.headline-page-button {
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  -ms-appearance: none;
-  -o-appearance: none;
-  appearance: none;
-  background-color: $gray-light;
-  border: none !important;
-  height: 12px;
-  width: 12px;
-  margin: 0 8px;
-  cursor: pointer;
-}
-
-.headline-page-button:checked {
-  background-color: $red;
-}
-
-.hot-tags-title {
-  font-weight: bold;
-  font-size: 16px;
-  line-height: 24px;
-  letter-spacing: 4px;
-}
-
-.headline-border {
-  border-left: 1px solid $gray-light;
-}
-
-.topic-recommend {
-  position: relative;
-  height: 424px;
-  margin: 0 -15px;
-
-  &::before {
-    content: '';
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    z-index: -1;
-    opacity: 0.4;
-    background-image: url('../../src/assets/images/TopicRecommendBG.svg');
+  @media screen and (max-width: 679px) {
+    margin-top: 2.5rem;
   }
+}
 
-  .title{
+.container_block {
+  display: block;
+}
+
+.section {
+  &_title {
+    font-size: 2rem;
     font-weight: 700;
-    font-size: 32px;
-    line-height: 40px;
+    line-height: 2.5rem;
     letter-spacing: 2px;
-  }
-
-  .content{
-    margin: 32px 0;
-    font-weight: 400;
-    font-size: 18px;
-    line-height: 24px;
+    margin-bottom: 2rem;
   }
 }
 
+.gallery {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 2rem;
+  margin: 0 auto;
+
+  &__wider {
+    width: 1444px;
+  }
+}
 </style>
